+++ GR_Datasets.lua
@@ -4,7 +4,6 @@ local _, ns = ... -- Namespace (myaddon, namespace)
 ns.ds = {}
 local ds = ns.ds
 
-ns.ds.GR_VERSION = '2.1.33' -- Show 'What's New' only if versions match
 function ds:Init()
     self.tblBadZones = ds:invalidZones()
     self.tblBadZonesByName = nil
@@ -15,32 +14,48 @@ function ds:Init()
     self.tblWhispers = {}
     self.tblConnected = ds:GetConnectedRealms()
 end
+ns.ds.GR_VERSION = '2.1.36' -- Show 'What's New' only if versions match
 function ds:WhatsNew()
     local update = false -- True and will save seen message
     local height = 410 -- Adjust size of what's new window
     local title, msg = '', ''
     title = ns.code:cText('FFFFFF00', "What's new in v"..GRADDON.version.."?")
     msg = [[
-        * Fixed invite issue with connected realms (Need feedback).
+         |CFF55d0ff** Please report any bugs or issues in Discord **
+                Discord: https://discord.gg/ZtS6Q2sKRH
+             (or click on the icon in the top left corner)|r
+
+    |CFFFFFF00v2.1.33 Notes|r
+        * Brought back adding black list players to the icon bar.
+        * Added the ability to send greeting/welcome message
+            when inviting via the menu (right click on
+            player name in chat).
+        * Added message length to greeting message and
+            restricted to 255 characters (1 message).
+        * Restricting welcome message to 255 characters.
+
+    |CFFFFFF00v2.1.33 Notes|r
+        * Fixed invite issue with connected realms.
         * Fixed anti-spam issue with connected realms.
         * UI improvements and clean up.
         * Default message is guild wide on your account.
-            Meaning, if you change it on one character, it will change on all.
+            Meaning, if you change it on one character, it will
+            change on all.
         * Added option to disable the 'What's New' message.
-        * Added skip if you don't want to invite a player right now,
-            it will add them to the skip list.
+        * Added skip if you don't want to invite a player right
+            now, it will add them to the skip list.
         * Scans now remember where you left off if you close
-            the addon and reopen (note: not if you log off or reload UI.)
+            the addon and reopen (note: not if you log off
+            or reload UI.)
         * Compact mode now remembers when you click on the icon.
         * Opened GM settings from any character on GM's account.
         * Added guild welcome message to the GM settings window.
-        * Added auto sync on login (will begin 60 seconds after login).
+        * Added auto sync on login (will begin 60 seconds
+            after login).
         * Fixed issues with auto sync not transferring all data.
 
         -> Sync with older versions will time out.
         -> Everyone needs to be on the current version.
-
-        ** Please report any bugs or issues on CurseForge or Discord **
     ]]
 
     return title, msg, height, update